requirements.txt
yt-dlp==2025.03.31
openai==1.70.0
python-dotenv
